+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Properly handle dynamic pool with zero minimum size.
+
 ## [3.1.13] - 2023-12-30
 
 ### Changed